Improve code styling and readability
