--- database/factories/ModelFactory.php
@@ -1,19 +1,0 @@
-<?php
-
-namespace AGILEDROP\LaravelTelnyx\Database\Factories;
-
-use Illuminate\Database\Eloquent\Factories\Factory;
-
-/*
-class ModelFactory extends Factory
-{
-    protected $model = YourModel::class;
-
-    public function definition()
-    {
-        return [
-
-        ];
-    }
-}
-*/
