+++ src/Channels/MmsChannel.php
@@ -2,21 +2,20 @@
 
 namespace AGILEDROP\LaravelTelnyx\Channels;
 
-use AGILEDROP\LaravelTelnyx\Messages\TelnyxMessage;
+use AGILEDROP\LaravelTelnyx\Messages\TelnyxMmsMessage;
 use Illuminate\Notifications\Notification;
+use Telnyx\Message;
 use Telnyx\Telnyx;
 
 class MmsChannel extends BaseChannel
 {
     /**
-     * Send the given notification trough MMS.
-     *
-     * The Telnyx API knows to send an MMS message when media_urls is specified.
+     * Send the given notification trough MMS
      *
      * @param  mixed  $notifiable
      * @param  \Illuminate\Notifications\Notification  $notification
      *
-     * @return \Telnyx\ApiResource|null
+     * @return \Telnyx\ApiResource|void
      */
     public function send($notifiable, Notification $notification)
     {
@@ -29,12 +28,12 @@ class MmsChannel extends BaseChannel
         $message = $notification->toTelnyx($notifiable);
 
         if (is_string($message)) {
-            $message = new TelnyxMessage($message);
+            $message = new TelnyxMmsMessage($message);
         }
 
         Telnyx::setApiKey(config('laravel-telnyx.api_key'));
 
-        return \Telnyx\Message::Create([
+        return Message::Create([
             "messaging_profile_id" => $this->profileId,
             'from' => $message->from ?: $this->from,
             'to' => $to,
