--- src/LaravelTelnyx.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace AGILEDROP\LaravelTelnyx;
-
-class LaravelTelnyx
-{
-}
